fix: use real object for concurrency tests
