+++ src/test/java/org/example/easytable/reservation/service/ReservationServiceTest.java
@@ -146,97 +146,4 @@ class ReservationServiceTest {
         // then
         verify(reservationRepository, times(1)).delete(reservation);
     }
-
-    @Test
-    @Transactional
-    public void checkReservationSaveConcurrency() throws InterruptedException {
-        // given
-        Long restaurantId = 1L;
-        int validSeatCount = 30;
-        int threadCount = 30;
-        int guestCount = 3;
-        Restaurant targetRestaurant = Restaurant.newRestaurant(
-                new RestaurantCreateDto("target", "addr1", validSeatCount));
-
-        ExecutorService executor = Executors.newFixedThreadPool(threadCount);
-        CountDownLatch startLatch = new CountDownLatch(1);
-        CountDownLatch doneLatch = new CountDownLatch(threadCount);
-        AtomicInteger successCnt = new AtomicInteger();
-        AtomicInteger failCnt = new AtomicInteger();
-
-        // when
-        restaurantRepository.save(targetRestaurant);
-
-        for (int i = 0; i < threadCount; i++) {
-            executor.submit(() -> {
-                try {
-                    startLatch.await();
-                    reservationService.save(restaurantId, LocalDateTime.now(), guestCount);
-                    successCnt.incrementAndGet();
-                } catch (Exception e) {
-                    System.out.println(e.getMessage());
-                    failCnt.incrementAndGet();
-                } finally {
-                    doneLatch.countDown();
-                }
-            });
-        }
-
-        startLatch.countDown();
-
-        doneLatch.await();
-        executor.shutdown();
-
-        // then
-        assertEquals(validSeatCount / guestCount, successCnt.intValue());
-        assertEquals(validSeatCount - (validSeatCount / guestCount), failCnt.intValue());
-    }
-
-    @Test
-    @Transactional
-    public void checkReservationDeleteConcurrency() throws InterruptedException {
-        // given
-        Long restaurantId = 1L;
-        int validSeatCount = 30;
-        int threadCount = 30;
-        int guestCount = 3;
-        Restaurant targetRestaurant = Restaurant.newRestaurant(
-            new RestaurantCreateDto("target", "addr1", validSeatCount));
-
-        ExecutorService executor = Executors.newFixedThreadPool(threadCount);
-        CountDownLatch startLatch = new CountDownLatch(1);
-        CountDownLatch doneLatch = new CountDownLatch(threadCount);
-        AtomicInteger successCnt = new AtomicInteger();
-        AtomicInteger failCnt = new AtomicInteger();
-
-        // when
-        restaurantRepository.save(targetRestaurant);
-
-        for (int i = 0; i < threadCount; i++) {
-            executor.submit(() -> {
-                try {
-                    startLatch.await();
-                    ReservationCreateResDto currentReservation = reservationService.save(
-                        restaurantId, LocalDateTime.now(), guestCount);
-                    reservationService.deleteReservation(restaurantId, currentReservation.getReservationId());
-                    successCnt.incrementAndGet();
-                } catch (Exception e) {
-                    System.out.println(e.getMessage());
-                    failCnt.incrementAndGet();
-                } finally {
-                    doneLatch.countDown();
-                }
-            });
-        }
-
-        startLatch.countDown();
-
-        doneLatch.await();
-        executor.shutdown();
-
-        // then
-        assertEquals(validSeatCount, targetRestaurant.getValidSeatCount());
-        assertEquals(threadCount, successCnt.intValue());
-        assertEquals(0, failCnt.intValue());
-    }
 }
